backend/requirements.txt
Django==5.2
psycopg[binary]==3.2.7 # For PostgreSQL connection (psycopg3)
celery==5.5
redis==5.0.7 # redis-py client, check TECH_STACK.md for Redis server version (8.0.0)
python-json-logger==3.3.0 # For structured logging
django-fernet-encrypted-fields==0.3.0 # For field encryption
django-allauth[socialaccount,openid]==65.7.0 # For Google OAuth
# Add gunicorn later for production, not needed for initial dev server
django-extensions==4.1
djangorestframework==3.16.0
dj-rest-auth==7.0.1
djangorestframework-simplejwt==5.5.0
django-celery-beat==2.8.0
croniter==6.0.0